requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-groq
pydantic
sqlalchemy
requests
google-auth
google-auth-oauthlib
google-api-python-client
python-multipart
python-multipart
